Moved protobuf to the geotrellis sub-package
