+++ geopyspark/geotrellis/protobufcodecs.py
@@ -7,13 +7,11 @@ check_environment()
 from geopyspark.geotrellis import (Extent, ProjectedExtent, TemporalProjectedExtent, SpatialKey,
                                    SpaceTimeKey)
 
-from geopyspark.protobuf.tileMessages_pb2 import ProtoTile, ProtoMultibandTile, ProtoCellType
-from geopyspark.protobuf.extentMessages_pb2 import (ProtoExtent, ProtoProjectedExtent,
-                                                    ProtoTemporalProjectedExtent)
-from geopyspark.protobuf import keyMessages_pb2
-from geopyspark.protobuf import extentMessages_pb2
-from geopyspark.protobuf import tileMessages_pb2
-from geopyspark.protobuf import tupleMessages_pb2
+from geopyspark.geotrellis.protobuf.tileMessages_pb2 import (ProtoTile, ProtoMultibandTile,
+                                                             ProtoCellType)
+from geopyspark.geotrellis.protobuf import keyMessages_pb2
+from geopyspark.geotrellis.protobuf import tupleMessages_pb2
+from geopyspark.geotrellis.protobuf import extentMessages_pb2
 
 
 _mapped_data_types = {
@@ -117,7 +115,7 @@ def extent_decoder(proto_bytes):
         :class:`~geopyspark.geotrellis.Extent`
     """
 
-    pb_extent = ProtoExtent.FromString(proto_bytes)
+    pb_extent = extentMessages_pb2.ProtoExtent.FromString(proto_bytes)
     return from_pb_extent(pb_extent)
 
 def from_pb_projected_extent(pb_projected_extent):
@@ -148,7 +146,7 @@ def projected_extent_decoder(proto_bytes):
         :class:`~geopyspark.geotrellis.ProjectedExtent`
     """
 
-    pb_projected_extent = ProtoProjectedExtent.FromString(proto_bytes)
+    pb_projected_extent = extentMessages_pb2.ProtoProjectedExtent.FromString(proto_bytes)
     return from_pb_projected_extent(pb_projected_extent)
 
 def from_pb_temporal_projected_extent(pb_temporal_projected_extent):
@@ -181,7 +179,7 @@ def temporal_projected_extent_decoder(proto_bytes):
         :class:`~geopyspark.geotrellis.TemporalProjectedExtent`
     """
 
-    pb_temporal_projected_extent = ProtoTemporalProjectedExtent.FromString(proto_bytes)
+    pb_temporal_projected_extent = extentMessages_pb2.ProtoTemporalProjectedExtent.FromString(proto_bytes)
     return from_pb_temporal_projected_extent(pb_temporal_projected_extent)
 
 def from_pb_spatial_key(pb_spatial_key):
