+++ geopyspark/tests/schema_tests/keys_schema_test.py
@@ -3,7 +3,7 @@ import pytest
 
 from pyspark import RDD
 from pyspark.serializers import AutoBatchedSerializer
-from geopyspark.protobuf import keyMessages_pb2
+from geopyspark.geotrellis.protobuf import keyMessages_pb2
 from geopyspark.geotrellis.protobufserializer import ProtoBufSerializer
 from geopyspark.geotrellis.protobufcodecs import (spatial_key_decoder,
                                                   spatial_key_encoder,
